+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-# v0.1.0 (2023-06-09)
+# v0.1.0 (2023-06-23)
 
 (Relative to [`bossy` v0.2.1](https://github.com/BrainiumLLC/bossy/blob/master/CHANGELOG.md#021-2021-01-08))